pyproject.toml
[project]
name = "foresight-r"
version = "0.1.0"
description = "Data science project for predictive analytics and modeling"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "scikit-learn>=1.4.0",
    "jupyter>=1.0.0",
    "ipython>=8.20.0",
    "plotly>=5.18.0",
    "scipy>=1.12.0",
    "loguru>=0.7.0",
    "hydra-core>=1.3.0",
    "pytest>=9.0.2",
    "meds-transforms==0.1.1",
    "es-aces>=0.6.2",
    "hydra-joblib-launcher>=1.2.0",
    "datasets>=4.0.0",
    "torch>=2.9.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "black>=24.0.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
]
ml = [
    "xgboost>=2.0.0",
    "lightgbm>=4.3.0",
    "optuna>=3.5.0",
]
deep-learning = [
    "torch>=2.2.0",
    "tensorflow>=2.15.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
custom-meds-transform-clean-string-columns = "foresight_r.meds.transforms.clean_string_columns:main"
custom-meds-transform-create-prefix-column = "foresight_r.meds.transforms.create_prefix_column:main"
custom-meds-transform-calculate-age = "foresight_r.meds.transforms.calculate_age:main"
custom-meds-transform-add-time-column = "foresight_r.meds.transforms.add_time_column:main"
custom-meds-transform-filter-null-values = "foresight_r.meds.transforms.filter_null_values:main"
custom-meds-transform-filter-by-text-value = "foresight_r.meds.transforms.filter_by_text_value:main"
custom-meds-transform-convert-to-markdown = "foresight_r.meds.transforms.convert_to_markdown:main"
custom-meds-transform-append-values = "foresight_r.meds.transforms.append_values:main"

custom-meds-transform-drop-by-regex = "foresight_r.meds.transforms.drop_by_regex:main"
custom-meds-transform-replace-code-with-text-value-by-prefix = "foresight_r.meds.transforms.replace_code_with_text_value_by_prefix:main"
custom-meds-transform-order-events = "foresight_r.meds.transforms.order_events:main"

validate-aces-prediction-times = "foresight_r.aces.validate_prediction_times:main"
combine-critical-outcome = "foresight_r.aces.combine_critical_outcome:main"

[tool.black]
line-length = 88
target-version = ['py313']

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
